Fixed compilation warnings [-Wsign-compare]
plugins/sailfish_wifi.c: In function ‘wifi_network_bss_update_strength’:
plugins/sailfish_wifi.c:1325:50: signed and unsigned type in conditional expression

src/service.c: In function ‘check_access’:
src/service.c:5333:13: comparison between signed and unsigned integer expressions
src/service.c:5353:13: comparison between signed and unsigned integer expressions

src/jolla-stats.c: In function ‘stats_delete_obsolete_files’:
src/jolla-stats.c:157:13: comparison between signed and unsigned integer expressions

+++ connman/plugins/sailfish_wifi.c
@@ -153,7 +153,7 @@ struct wifi_bss {
 	gulong event_id[BSS_EVENT_COUNT];
 	guint remove_timeout_id;
 	GUtilIntHistory *history;       /* Signal strength history */
-	int strength;                   /* Median strength */
+	guint strength;                 /* Median strength */
 };
 
 typedef enum wifi_network_state {
@@ -1322,7 +1322,7 @@ static void wifi_network_bss_update_strength(gpointer data, gpointer user_data)
 	struct wifi_bss *bss_data = data;
 
 	bss_data->strength = gutil_int_history_size(bss_data->history) ?
-		gutil_int_history_median(bss_data->history, 0) :
+		(guint)gutil_int_history_median(bss_data->history, 0) :
 		wifi_rssi_strength(bss_data->bss->signal);
 }
